Redone the conflict detection code to almost directly copy the Minecraft detection. Thanks to DrMeteor for spotting some discrepancies between last commit detection and MC's detection.
